dhcp: allow a dhcp-timeout value of -1 to disable dhcp timeout
-1 will then be converted to G_MAXINT32.
Then exact behavior depends on the selected dhcp client:

dhclient: G_MAXINT32 will be used as argument to the --timeout
parameter. The timeout will last roughly 68 years, so there will be no
timeout in practice.

internal: G_MAXINT32 will be recognised as no timeout and no timer will
be started.

dhcpd: we do not support passing a custom timeout value here, neither we
are going to support disabling it.